third_party/blink/web_tests/external/wpt/css/css-overflow/scroll-markers-focus-on-scrolling-expected.txt

This is a testharness.js-based test.
[FAIL] active ::scroll-marker saves focus on scroll if previous ::scroll-marker had it
  assert_equals: ::scroll-marker opacity should be 0.5 but was 1 expected "0.5" but got "1"
Harness: the test ran to completion.
